Move HistoryModel and RowData to the new history module
Feels like a better organization
