+++ src/history.rs
@@ -1,3 +1,7 @@
 mod context_menu;
+mod model;
+mod row_data;
 
 pub use context_menu::context_menu;
+pub use model::{HistoryModel, ModelNotifier};
+pub use row_data::RowData;
