+++ src/main.rs
@@ -14,7 +14,6 @@ mod appwindow;
 mod db;
 mod editor;
 mod history;
-mod historymodel;
 mod kcshot;
 mod postcapture;
 mod settings_window;
